refactor: extract app bootstrap and shutdown orchestration from Mode-S Client
